requirements.txt
certifi==2021.5.30
cycler==0.11.0
imutils==0.5.4
joblib==1.1.1
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.19.5
opencv-contrib-python==3.4.2.16
opencv-python==3.4.2.16
pandas==1.1.5
Pillow==8.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
threadpoolctl==3.1.0
